--- a/.github/workflows/coverage.yml
+++ b/.github/workflows/coverage.yml
@@ -42,4 +42,9 @@ jobs:
       - name: Upload to codecov.io
         uses: codecov/codecov-action@v4
         with:
+          fail_ci_if_error: true # optional (default = false)
           files: coverage/*.lcov
+          token: ${{ secrets.CODECOV_TOKEN }} # required
+          verbose: true # optional (default = false)
+          name: ractor # optional
+